Refactor code and + task
